update Makefile
In production server, we do not set password for root user of
mysql.
So, we need to choose the way executing mysql command as root
according to existence of root user password environment variable.

Actually, the code in this Makefile is ugly..
I think we should refactor it some day..